pyproject.toml
[project]
name = "dakoda-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars (>=1.30.0,<2.0.0)",
    "dkpro-cassis (>=0.10.1,<1.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "xsdata (>=25.4,<26.0.0)",
]

[tool.poetry]
packages = [
    { include = "dakoda"},
    { include = "dakoda_typesystem.xml", from = "dakoda/res" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]